requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
serpapi==0.1.5
python-dotenv==1.0.0
yt-dlp==2023.10.13
langchain==0.0.346
pydantic==2.5.0